+++ src/main/java/com/example/springboot/service/UserService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface UserService {
     void add(User user);
-    User getById(int id);
+    User getById(long id);
     User getByName(String name);
     void updateUser(User user);
     List<User> getAllUsers();
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
